+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2025-08-23
+
+### Added
+
+- New `st_smart` table created to store `smartctl` scan results.
+- `scan_drive.php` now saves `smartctl` output for each drive scan into the `st_smart` table.
+- "Smartctl" action added to `drives.php` to display `smartctl` data for a drive.
+
 ## [1.0.4] - 2025-08-23
 
 ### Added